Add mixins for browser prefixed transitions & grayscale
Adds a mixin to use for cross browser transition prefixes:
transitions(). This mixin takes the $properties to transition (as a
list), the transition $duration, the transition $timing-function, and
the transition $delay as arguments.

I'd like to get this working in outputting the CSS transition shorthand,
but Sass doesn't seem to be able to loop within the property's value.
Perhaps if I push values onto the end of a temporary array...

This commit also adds the saturation() mixin. This mixin takes the
$amount of saturation as an argument. It should enable grayscale for a
bunch of browsers: Chrome 18+, Safari 6+, Opera 15+, iOS Safari 6+,
Android Browser 4.4+, Opera Mobile 24+, Blackberry Browser 10+, Chrome
for Android 41+, Firefox for Android 36+, current versions of any Webkit
based browser, UC Browser for Android 9.9+, IE 5.5+, and spec support.

This commit adds the fadein animation for use when fading objects into
view.