Use jQuery instead of d3
D3.js was being used here to manipulate a few tile elements, but no
charting was being done. That doesn't sound like a great reason to load
another library to me. Let's only load d3 when we need it to make some
data driven documents.

We are already loading jQuery for other dependencies, so lets just use
that for now.

Also, the inline javascript in the blog post has been moved into the
javascript folder for ease of maintanance (and possible reuse). The
javascript has been re-implemented in jQuery.