Move SharePoint tile markup into partial
This commit moves the html that was being used to generate the mock
SharePoint tiles into a partial to promote reuse, easier maintenance,
and to support the javascript module approach.

The blog post now needs to be parsed by erb before being run through the
markdown parser (the partial injection was troublesome with haml).

A folder was created to house all of the partials (called "partials") to
keep things clean and organized.