Clean up background images
Darken the images themselves so that the title text stands out a bit
more. Also, rather than dumping styles directly into the layouts, insert
a background image using JavaScript based off of the data image
attribute.

This way, we can use the same bit of code to handle the homepage, work,
and blog background images.

I've added the ability to direct where the background image will go and
pass it options (like whether or not the image should be inserted in
mobile devices).